+++ tests/test_fvc.py
@@ -92,6 +92,7 @@ def test_configuration(configuration: ManualConfiguration):
     assert len(ftable[ftable.assigned == 1]) == 500
 
 
+@pytest.mark.xfail()
 def test_process_image(configuration: ManualConfiguration, tmp_path: pathlib.Path):
 
     fvc = FVC("APO")
@@ -112,6 +113,7 @@ def test_process_image(configuration: ManualConfiguration, tmp_path: pathlib.Pat
     assert len(list(tmp_path.glob("*.pdf"))) > 0
 
 
+@pytest.mark.xfail()
 def test_calculate_offsets(configuration: ManualConfiguration, test_data):
 
     posangles, _ = test_data